\name{murl-package}
\Rdversion{1.1}
\alias{murl-package}
\alias{murl}
\docType{package}
\title{Mailmerge methods for reading addresses, creating letters, and mapping US ZIP codes of recipients}

\description{
Provides mailmerge methods for reading spreadsheets of addresses and other relevant information to create standardized but customizable letters.  Provides a method to map US ZIP codes, including those of letter recipients.  Provides a method for parsing and processing html code from online job postings of the American Political Science Association.
}

\details{
\tabular{ll}{
Package: \tab muRL\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2009-07-31\cr
License: \tab GPL version 2.0 or newer \cr
URL: \tab \url{http://rtm.wustl.edu/software.muRL.htm} \cr
LazyLoad: \tab yes\cr
}

%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}

\author{
Ryan T. Moore \email{rtm@wustl.edu} and Andrew Reeves \email{areeves@bu.edu}

Maintainer: Ryan T. Moore <rtm@wustl.edu>
}

%\references{}

\keyword{ package }

%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}

%\examples{
%~~ simple examples of the most important functions ~~
%}